Resolve issue #1
Resolved issue #1

In line 40, the argument to execGRASS should have been the filename in r_layer. However, we were passing layer. We got away with it for a while because layer is the name of the function arg, and it is meant to be a file path, so if you passed layer as a path the function still worked.

The current solution now works for path and raster layer input.